+++ api/api.js
@@ -1,1 +1,7 @@
+const config = require('./../config').api;
+
 const KrakenClient = require('kraken-api');
+const key = config.key; // API Key
+const secret = config.secret; // API Private Key
+const kraken = new KrakenClient(key, secret);
+
